requirements_full.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.0.3
python-dateutil==2.8.2
pydantic==2.5.3
python-multipart==0.0.6
aiohttp==3.9.3  # For testing connection pool
sentence-transformers>=2.3.0
torch>=1.9.0
numpy>=1.21.0,<2.0.0
slowapi==0.1.9  # For rate limiting
pytest-asyncio==1.0.0  # For async tests